Example with verifying commit by content
It's basically the same as the commit-message-length-validator.jar,
but written in groovy and using the contents of the commit.

This version checks all added lines for "DO_NOT_COMMIT" as I add
that comment on/around all lines where I add stuff (e.g. debug-code)
that shouldn't reach the repo.